# workspace/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope_ContentGenerationResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentGenerationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ApiResponseStatus"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "success",
          "status",
          "data",
          "error",
          "timestamp"
        ],
        "title": "ApiEnvelope[ContentGenerationResponse]",
        "type": "object"
      },
...
      "ApiResponseStatus": {
        "description": "API request processing status.",
        "enum": [
          "SUCCESS",
          "PROCESSING",
          "FAILED"
        ],
        "title": "ApiResponseStatus",
        "type": "string"
      },
...
      "ContentPurpose": {
        "description": "User-facing purpose that selects the copy generation rules.",
        "enum": [
          "INTRODUCTION",
          "NEWS"
        ],
        "title": "ContentPurpose",
        "type": "string"
      },
      "CreateContentGenerationRequest": {
        "additionalProperties": false,
        "description": "Request to create one three-platform content generation.",
        "properties": {
          "briefText": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Brieftext",
            "type": "string"
          },
          "purpose": {
            "$ref": "#/components/schemas/ContentPurpose",
            "default": "INTRODUCTION"
          },
          "seedKeywords": {
            "items": {
              "maxLength": 30,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 0,
            "title": "Seedkeywords",
            "type": "array"
          },
          "sourceReviewIds": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 10,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcereviewids"
          },
          "storeProfileId": {
            "format": "uuid",
            "title": "Storeprofileid",
            "type": "string"
          },
          "toneInstruction": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toneinstruction"
          }
        },
        "required": [
          "briefText",
          "seedKeywords",
          "storeProfileId"
        ],
        "title": "CreateContentGenerationRequest",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Failure envelope: ``success`` is false, ``data`` is null and ``error`` is set.",
        "properties": {
          "data": {
            "title": "Data",
            "type": "null"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ApiResponseStatus"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "success",
          "status",
          "data",
          "error",
          "timestamp"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "MapKeeper MVP v0.2. JSON fields are camelCase, ids are UUID strings and timestamps are UTC ISO 8601. Undefined request and response fields are rejected.",
    "title": "MapKeeper API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/seo/generations": {
      "post": {
        "description": "Take one common input and produce exactly one result per platform.",
        "operationId": "create_generation_api_v1_seo_generations_post",
        "parameters": [
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ContentGenerationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate one result for each of the three platforms",
        "tags": [
          "seo"
        ]
      }
    },
    "/api/v1/seo/generations/{generationId}/approve": {
      "post": {
        "description": "Commit whole-generation approval before queuing synchronization.",
        "operationId": "approve_generation_api_v1_seo_generations__generationId__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "generationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Generationid",
              "type": "string"
            }
          },
          {
            "description": "Required on every approval. Same actor and key with the same approval target replays the existing SyncJob; a different target is 409 IDEMPOTENCY_CONFLICT.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Required on every approval. Same actor and key with the same approval target replays the existing SyncJob; a different target is 409 IDEMPOTENCY_CONFLICT.",
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ContentGenerationApprovalResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No resource matches the requested id. error.code is one of: RESOURCE_NOT_FOUND."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested transition cannot be performed in the current state. error.code is one of: INVALID_STATE, STALE_PROPOSAL, IDEMPOTENCY_CONFLICT, NO_RETRYABLE_TASKS."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve a whole DRAFT generation and start synchronization",
        "tags": [
          "seo"
        ]
      }
    },
    "/api/v1/seo/generations/{generationId}/drafts": {
      "patch": {
        "description": "Store the owner's corrections so approval publishes what they read.",
        "operationId": "edit_generation_drafts_api_v1_seo_generations__generationId__drafts_patch",
        "parameters": [
          {
            "in": "path",
            "name": "generationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Generationid",
              "type": "string"
            }
          },
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditContentDraftsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ContentGenerationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No resource matches the requested id. error.code is one of: RESOURCE_NOT_FOUND."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested transition cannot be performed in the current state. error.code is one of: INVALID_STATE, STALE_PROPOSAL, IDEMPOTENCY_CONFLICT, NO_RETRYABLE_TASKS."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replace the copy of a DRAFT generation with the owner's edits",
        "tags": [
          "seo"
        ]
      }
    },
    "/api/v1/seo/generations/{generationId}/regenerate": {
      "post": {
        "description": "Replace all platform results and increment revision on a DRAFT.",
        "operationId": "regenerate_generation_api_v1_seo_generations__generationId__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "generationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Generationid",
              "type": "string"
            }
          },
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateContentGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ContentGenerationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No resource matches the requested id. error.code is one of: RESOURCE_NOT_FOUND."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested transition cannot be performed in the current state. error.code is one of: INVALID_STATE, STALE_PROPOSAL, IDEMPOTENCY_CONFLICT, NO_RETRYABLE_TASKS."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replace all three results of a DRAFT generation",
        "tags": [
          "seo"
        ]
      }
    },
    "/api/v1/seo/generations/{generationId}/reject": {
      "post": {
        "description": "Move the whole generation to REJECTED. The request has no body.",
        "operationId": "reject_generation_api_v1_seo_generations__generationId__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "generationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Generationid",
              "type": "string"
            }
          },
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ContentGenerationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No resource matches the requested id. error.code is one of: RESOURCE_NOT_FOUND."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested transition cannot be performed in the current state. error.code is one of: INVALID_STATE, STALE_PROPOSAL, IDEMPOTENCY_CONFLICT, NO_RETRYABLE_TASKS."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject a whole DRAFT generation",
        "tags": [
          "seo"
        ]
      }
    },
...
    "/api/v1/sync-jobs/{syncJobId}": {
      "get": {
        "description": "Report job progress. A failed platform still returns 200 with status SUCCESS.",
        "operationId": "get_sync_job_api_v1_sync_jobs__syncJobId__get",
        "parameters": [
          {
            "in": "path",
            "name": "syncJobId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Syncjobid",
              "type": "string"
            }
          },
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SyncJobStatusResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No resource matches the requested id. error.code is one of: RESOURCE_NOT_FOUND."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read the aggregate job status and per-platform detail",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}